Refactor ConsulBasedTargetProvider to allow for custom load balancing strategies.
