+++ ob1k-consul/src/main/java/com/outbrain/ob1k/consul/ConsulBasedTargetProvider.java
@@ -1,22 +1,21 @@
 package com.outbrain.ob1k.consul;
 
-import com.google.common.base.Preconditions;
-import com.outbrain.ob1k.client.targets.TargetProvider;
-import io.netty.util.internal.ThreadLocalRandom;
-import org.apache.commons.lang3.StringUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import static com.google.common.base.Preconditions.checkArgument;
+import static com.google.common.base.Preconditions.checkState;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import static com.google.common.base.Preconditions.checkArgument;
-import static com.google.common.base.Preconditions.checkState;
+import org.apache.commons.lang3.StringUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import com.google.common.base.Preconditions;
+import com.outbrain.ob1k.client.targets.TargetProvider;
 
 /**
  * A {@link TargetProvider} that provides targets registered in consul.
@@ -27,23 +26,22 @@ public class ConsulBasedTargetProvider implements TargetProvider, HealthyTargets
 
   private static final Logger log = LoggerFactory.getLogger(ConsulBasedTargetProvider.class);
 
-  private final ThreadLocal<Integer> currIndex = new ThreadLocal<Integer>() {
-    @Override
-    protected Integer initialValue() {
-      return ThreadLocalRandom.current().nextInt();
-    }
-  };
-
   private final String urlSuffix;
   private final Map<String, Integer> tag2weight;
   private final HealthyTargetsList healthyTargetsList;
   private volatile List<String> targets;
+  private final LoadBalancer loadBalancer;
 
   public ConsulBasedTargetProvider(final HealthyTargetsList healthyTargetsList, final String urlSuffix, final Map<String, Integer> tag2weight) {
+    this(healthyTargetsList, urlSuffix, tag2weight, new ThreadLocalRoundRobinLoadBalancer());
+  }
+
+  public ConsulBasedTargetProvider(final HealthyTargetsList healthyTargetsList, final String urlSuffix, final Map<String, Integer> tag2weight, LoadBalancer loadBalancer) {
     this.urlSuffix = urlSuffix == null ? "" : urlSuffix;
     this.tag2weight = tag2weight == null ? Collections.<String, Integer>emptyMap() : new HashMap<>(tag2weight);
     this.healthyTargetsList = Preconditions.checkNotNull(healthyTargetsList, "healthyTargetsList must not be null");
     healthyTargetsList.addListener(this);
+    this.loadBalancer = loadBalancer;
   }
 
   @Override
@@ -73,17 +71,7 @@ public class ConsulBasedTargetProvider implements TargetProvider, HealthyTargets
     checkState(!currTargets.isEmpty(), "No targets are currently registered for module " + healthyTargetsList.getModule());
     checkArgument(targetsNum > 0, "targets number must be more than zero");
 
-    final int targetsSize = currTargets.size();
-    final int index = currIndex.get();
-
-    currIndex.set(index + 1);
-
-    final List<String> providedTargets = new LinkedList<>();
-    for (int i = 0; i < targetsNum; i++) {
-      providedTargets.add(currTargets.get(Math.abs((index + i) % targetsSize)));
-    }
-
-    return providedTargets;
+    return loadBalancer.provideTargets(currTargets, targetsNum);
   }
 
   @Override
@@ -99,6 +87,7 @@ public class ConsulBasedTargetProvider implements TargetProvider, HealthyTargets
 
     Collections.shuffle(targets);
     this.targets = targets;
+    loadBalancer.onTargetsChanged(targets);
     log.debug("New weighed targets: {}", targets);
   }
