+++ ob1k-consul/src/main/java/com/outbrain/ob1k/consul/LoadBalancer.java
@@ -1,0 +1,10 @@
+package com.outbrain.ob1k.consul;
+
+import java.util.List;
+
+public interface LoadBalancer {
+
+  List<String> provideTargets(final List<String> targets, final int targetsNum);
+
+  default void onTargetsChanged(final List<String> newTargets) {}
+}
